RC Release (Stable and useful code)
Many improvements. It's finally a working project.

Now you can create a "Mac", which is basically a folder with configuration file and optionally a hard disk file (.dsk).

Also you can edit the config via a GUI. It needs some minor tweaks (and better images), but it works for the most.

Note that any comments made in previously made config files (or any config files) won't be saved.

This is a stable RC. You can use it to run and save several PearPC configurations, but as mentioned, it needs some tweaks to be complete.